Fix return format of X- and R-Learner; Fix default method used
The return format of X- and R-Learner was incompatible with what the
metrics in justcause.metrics expected, because of the nested array
format. Thus, the outputs are flattened, which works fine as long as
only one treatment is considered.

The default method of the RLearner is changed to one that allows the use
of sample_weights in the fit() method.