Revert "Use generators in eval; Use dict over np.array"
This reverts commit 837152a93e2b985a2644823b6d7d09ff6d284158.
Using generators within the automatic evaluation (evaluate_ite) results
in wrong usage of replications, as the generator cannot be reset between
methods. Thus, different replications would be used, which is
dangerous!
For manual evaluation loops, it is still recommended to work with
generators in order to avoid clogging the memory.